move voxel traversal algorithm functions to moudle vta
